Move number blueprint to number module
