+++ src/yafowil/tests/__init__.py
@@ -4,6 +4,7 @@ from node.tests import NodeTestCase
 from yafowil import common
 from yafowil import compound
 from yafowil import datatypes
+from yafowil import number
 from yafowil import persistence
 from yafowil import structure
 from yafowil import table
@@ -23,6 +24,7 @@ class YafowilTestCase(NodeTestCase):
         reload(common)
         reload(compound)
         reload(datatypes)
+        reload(number)
         reload(persistence)
         reload(structure)
         reload(table)
@@ -50,6 +52,7 @@ def test_suite():
     from yafowil.tests import test_compound
     from yafowil.tests import test_controller
     from yafowil.tests import test_datatypes
+    from yafowil.tests import test_number
     from yafowil.tests import test_persistence
     from yafowil.tests import test_resources
     from yafowil.tests import test_structure
@@ -64,6 +67,7 @@ def test_suite():
     suite.addTest(unittest.findTestCases(test_compound))
     suite.addTest(unittest.findTestCases(test_controller))
     suite.addTest(unittest.findTestCases(test_datatypes))
+    suite.addTest(unittest.findTestCases(test_number))
     suite.addTest(unittest.findTestCases(test_persistence))
     suite.addTest(unittest.findTestCases(test_resources))
     suite.addTest(unittest.findTestCases(test_structure))
